Upgrade dokuwiki to 2020-07-29 "Hogfather" and upgrade plugins
This commit upgrade the docker container with latest available
dokuwiki version and upgrades all plugins versions.

It also adds the authsaml plugin + simplesamlphp (which is a
dependency of the pluging).

Both the plugin and simplesamlphp needs to be configured, please
check the README of the plugin for configuration:
https://github.com/pitbulk/dokuwiki-saml

In order to persist both config we advise you to mount a volume to
both directories:
- /var/www/lib/plugins/authsaml
- /var/www/simplesamlphp